requirements.txt
pandas
sqlalchemy
requests
bs4
airflow
dotenv
psycopg2-binary